haidz123

+++ src/main/java/hotel/HotelServlet.java
@@ -8,15 +8,14 @@ import jakarta.servlet.http.*;
 
 import javax.annotation.Resource;
 import javax.sql.DataSource;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
+import java.io.*;
 import java.sql.SQLException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
+
+import static java.lang.System.out;
 
 @WebServlet("/HotelServlet")
 @MultipartConfig
@@ -121,6 +120,43 @@ public class HotelServlet extends HttpServlet {
         return "";
     }
 
+
+//    private void bookRoom(HttpServletRequest request, HttpServletResponse response, int roomId) throws ServletException, IOException {
+//        try {
+//            HttpSession session = request.getSession();
+//            String userId = (String) session.getAttribute("userId");
+//            if (userId != null) {
+//                String checkInDate = request.getParameter("checkInDate");
+//                String checkOutDate = request.getParameter("checkOutDate");
+//                int hotelId = Integer.parseInt(request.getParameter("hotel_id"));
+//                double totalPrice = Double.parseDouble(request.getParameter("totalPrice"));
+//                String status = "PENDING";
+//
+//                // Kiểm tra xem phòng có sẵn trong khoảng thời gian đặt hay không
+//                if (hotelDB.isRoomAvailable(roomId, checkInDate, checkOutDate)) {
+//                    Booking booking = new Booking(Integer.parseInt(userId), roomId, hotelId, checkInDate, checkOutDate, totalPrice, status);
+//                    hotelDB.addBooking(booking);
+//
+//                    hotelDB.deleteCartItemsByUserId(Integer.parseInt(userId), roomId);
+//                     // Sau khi kiểm tra phòng khả dụng và hoàn tất quá trình đặt phòng
+//                    hotelDB.updateRoomAvailability(roomId, false); // Đặt trạng thái phòng không khả dụng
+//
+//                    response.sendRedirect(request.getContextPath() + "/HotelServlet?command=LOAD");
+//                } else {
+//                    // Phòng không khả dụng, chuyển hướng về trang thông báo
+//                    response.sendRedirect(request.getContextPath() + "/unavailable.jsp");
+//                }
+//            } else {
+//                // Redirect to login page or display an error message
+//                response.sendRedirect(request.getContextPath() + "/login.jsp");
+//            }
+//        } catch (NumberFormatException e) {
+//            throw new ServletException("Invalid number format", e);
+//        } catch (Exception e) {
+//            throw new ServletException(e);
+//        }
+//    }
+
     private void bookRoom(HttpServletRequest request, HttpServletResponse response, int roomId) throws ServletException, IOException {
         try {
             HttpSession session = request.getSession();
@@ -128,15 +164,34 @@ public class HotelServlet extends HttpServlet {
             if (userId != null) {
                 String checkInDate = request.getParameter("checkInDate");
                 String checkOutDate = request.getParameter("checkOutDate");
-                double totalPrice = Double.parseDouble(request.getParameter("totalPrice"));
-                String status = "PENDING";
+                int hotelId = Integer.parseInt(request.getParameter("hotel_id"));
+                int adults = Integer.parseInt(request.getParameter("adults"));
+                int children = Integer.parseInt(request.getParameter("children"));
+                String email = request.getParameter("email");
+                String name = request.getParameter("name");
+                String phone = request.getParameter("phone");
 
-                Booking booking = new Booking(Integer.parseInt(userId), roomId, checkInDate, checkOutDate, totalPrice, status);
-                hotelDB.addBooking(booking);
+                // Kiểm tra tính khả dụng của phòng
+                boolean isRoomAvailable = hotelDB.isRoomAvailable(roomId, checkInDate, checkOutDate);
 
-                hotelDB.deleteCartItemsByUserId(Integer.parseInt(userId),roomId);
+                if (isRoomAvailable) {
+                    double totalPrice = Double.parseDouble(request.getParameter("totalPrice"));
+                    String status = "PENDING";
+                    Booking booking = new Booking(Integer.parseInt(userId), roomId, hotelId, checkInDate, checkOutDate, totalPrice, status);
+                    booking.setAdults(adults);
+                    booking.setChildren(children);
+                    int bookingId = hotelDB.addBookings(booking); // Lưu thông tin vào bảng bookings và lấy booking_id
+//                    hotelDB.addBooking(booking);
+                    hotelDB.saveCustomer(userId, name, phone, email, bookingId);
+                    hotelDB.deleteCartItemsByUserId(Integer.parseInt(userId), roomId);
+                    hotelDB.updateRoomAvailability(roomId, false); // Đặt trạng thái phòng không khả dụng
+                    response.sendRedirect(request.getContextPath() + "/complete.jsp");
 
-                response.sendRedirect(request.getContextPath() + "/HotelServlet?command=LOAD");
+//                    response.sendRedirect(request.getContextPath() + "/HotelServlet?command=LOAD");
+                } else {
+                    // Phòng không khả dụng, chuyển hướng về trang thông báo
+                    response.sendRedirect(request.getContextPath() + "/unavailable.jsp");
+                }
             } else {
                 // Redirect to login page or display an error message
                 response.sendRedirect(request.getContextPath() + "/login.jsp");
@@ -148,6 +203,14 @@ public class HotelServlet extends HttpServlet {
         }
     }
 
+    private double calculateTotalPrice(HttpServletRequest request, int adults, int children) {
+        double priceA = 30; // Giá A
+        double priceB = 50; // Giá B
+
+        double totalPrice = (priceA * adults) + (priceB * children);
+
+        return totalPrice;
+    }
 
 
 
@@ -159,21 +222,36 @@ public class HotelServlet extends HttpServlet {
 
 
 
-
     private void addToCart(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         try {
             HttpSession session = request.getSession();
             String userId = (String) session.getAttribute("userId");
             if (userId != null) {
                 int roomId = Integer.parseInt(request.getParameter("roomId"));
-                hotelDB.saveRoomToCart(userId, roomId);
-            }
-            else {
+
+                // Kiểm tra tính khả dụng của availability
+                int availability = hotelDB.getRoomAvailability(roomId);
+                if (availability != 0) {
+
+                    if (hotelDB.isRoomInCart(userId, roomId)) {
+                        RequestDispatcher dispatcher = request.getRequestDispatcher("/roomincart.jsp");
+                        dispatcher.forward(request, response);
+                        return;
+                    }
+
+                    hotelDB.saveRoomToCart(userId, roomId);
+                    RequestDispatcher dispatcher = request.getRequestDispatcher("/available.jsp");
+                    dispatcher.forward(request, response);
+                } else {
+                    RequestDispatcher dispatcher = request.getRequestDispatcher("/unavailable.jsp");
+                    dispatcher.forward(request, response);
+                }
+            } else {
                 // Redirect to login page or display an error message
                 response.sendRedirect(request.getContextPath() + "/login.jsp");
             }
-            String currentURL = request.getRequestURL().toString();
 
+            String currentURL = request.getRequestURL().toString();
             // Chuyển hướng người dùng đến chính trang hiện tại
             response.sendRedirect(currentURL);
 
@@ -221,24 +299,39 @@ public class HotelServlet extends HttpServlet {
 
         // Lấy danh sách thông tin khách hàng (full_name và email) dựa trên user_id
         List<String> fullNames = new ArrayList<>();
-        List<String> rooms = new ArrayList<>();
+        List<String> price1 = new ArrayList<>();
         List<String> emails = new ArrayList<>();
+        List<String> hotelNames = new ArrayList<>();
+        List<Customers> customers = new ArrayList<>();
+
 
         for (Booking booking : bookings) {
             String fullName = hotelDB.getFullName(booking.getUser_id());
             String email = hotelDB.getEmail(booking.getUser_id());
             Double price = Double.parseDouble(hotelDB.getPrice(booking.getRoom_id()));
+            String hotelName = hotelDB.getHotelName(booking.getHotel_id());
 
             fullNames.add(fullName);
             emails.add(email);
-            rooms.add(String.valueOf(price));
+            price1.add(String.valueOf(price));
+            hotelNames.add(hotelName);
+            Customers customer = hotelDB.getCustomerById(booking.getUser_id());
+            customers.add(customer);
 
+            // Lấy thông tin từ bảng hotels
+            int hotelId = booking.getHotel_id();
+            Hotel hotel = hotelDB.getHotelById(hotelId);
+//            booking.setHotelName(hotel.getHotel_name());
+            booking.setHotel_id(booking.getHotel_id());
         }
 
         request.setAttribute("BOOKING_LIST", bookings);
         request.setAttribute("FULL_NAMES", fullNames);
         request.setAttribute("EMAILS", emails);
-        request.setAttribute("PRICE", rooms);
+        request.setAttribute("PRICE", price1);
+        request.setAttribute("HOTEL_NAMES", hotelNames);
+        request.setAttribute("CUSTOMERS", customers);
+
 
         // Forward to the booking-list.jsp
         RequestDispatcher dispatcher = request.getRequestDispatcher("/booking-list.jsp");
@@ -246,18 +339,6 @@ public class HotelServlet extends HttpServlet {
     }
 
 
-//    private void loadBookingDetails(HttpServletRequest request, HttpServletResponse response) throws Exception {
-//        HttpSession session = request.getSession();
-//        String userId = (String) session.getAttribute("userId");
-//
-//
-//        List<Booking> bookings = hotelDB.getBookings(Integer.parseInt(userId));
-//
-//        request.setAttribute("BOOKING_LIST", bookings);
-//        // Forward to the hotel-details.jsp
-//        RequestDispatcher dispatcher = request.getRequestDispatcher("/booking-list.jsp");
-//        dispatcher.forward(request, response);
-//    }
 
 
     private void loadCartDetails(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
@@ -282,6 +363,9 @@ public class HotelServlet extends HttpServlet {
                     int hotelId = room.getHotel_id();
                     Hotel hotel = hotelDB.getHotelById(hotelId);
                     cartItem.setHotelName(hotel.getHotel_name());
+
+                    cartItem.setHotel_id(room.getHotel_id());
+
 
                     // Lấy thông tin từ bảng users (từ session)
                     cartItem.setUserId(Integer.parseInt(userId));
@@ -349,13 +433,13 @@ public class HotelServlet extends HttpServlet {
             HttpSession session = request.getSession();
             session.setAttribute("userId", String.valueOf(userId));
 
-            System.out.println("userID : " + userId);
+            out.println("userID : " + userId);
 
 
             // Tiến hành các hành động tiếp theo
             response.sendRedirect(request.getContextPath() + "/HotelServlet");
         } else {
-            System.out.println("bú");
+            out.println("bú");
         }
     }
